+++ alacritty/src/event.rs
@@ -440,10 +440,10 @@ impl ApplicationHandler<Event> for Processor {
                 }
             },
             // NOTE: This event bypasses batching to minimize input latency.
-            (EventType::Frame, Some(window_id)) => {
+            (EventType::Frame { force }, Some(window_id)) => {
                 if let Some(window_context) = self.windows.get_mut(window_id) {
                     window_context.display.window.has_frame = true;
-                    if window_context.dirty {
+                    if window_context.dirty || force {
                         window_context.display.window.request_redraw();
                     }
                 }
@@ -555,7 +555,9 @@ pub enum EventType {
     SearchNext,
     #[cfg(unix)]
     Shutdown,
-    Frame,
+    Frame {
+        force: bool,
+    },
 }
 
 impl From<TerminalEvent> for EventType {
@@ -774,7 +776,9 @@ impl<'a, N: Notify + 'a, T: EventListener> input::ActionContext<T> for ActionCon
         point.line = min(point.line, self.terminal.bottommost_line());
 
         // Update selection.
-        selection.update(point, side);
+        if !self.display().scrollbar.is_dragging() {
+            selection.update(point, side);
+        }
 
         // Move vi cursor and expand selection.
         if self.terminal.mode().contains(TermMode::VI) && !self.search_active() {
@@ -1713,6 +1717,7 @@ pub enum TouchPurpose {
     None,
     Select(TouchEvent),
     Scroll(TouchEvent),
+    ScrollbarDrag(TouchEvent),
     Zoom(TouchZoom),
     ZoomPendingSlot(TouchEvent),
     Tap(TouchEvent),
@@ -1933,7 +1938,7 @@ impl input::Processor<EventProxy, ActionContext<'_, Notifier, EventProxy>> {
                 EventType::Message(_)
                 | EventType::ConfigReload(_)
                 | EventType::CreateWindow(_)
-                | EventType::Frame => (),
+                | EventType::Frame { .. } => (),
             },
             WinitEvent::WindowEvent { event, .. } => {
                 match event {