+++ CHANGELOG.md
@@ -62,6 +62,7 @@ Notable changes to the `alacritty_terminal` crate are documented in its
 
 ### Added
 
+- Scrollbar (enable via config `scrollbar.mode`)
 - Vi motions `*`, `#`, `{`, and `}`
 - IPC config retrieval using `alacritty msg get-config`
 - Multi-sequence touch zoom sequences